Update CopyRight - minor fixes
