+++ .travis.yml
@@ -15,6 +15,6 @@ install:
   - "pip install ."
   - "pip install -r requirements.txt"
 script:
-  - nosetests --with-coverage --cover-package=pyensemblrest tests/alltests.py
+  - nosetests --with-coverage --cover-package=pyensemblrest tests/test_all.py
 after_success:
   - coveralls
