+++ TODO
@@ -1,0 +1,8 @@
+
+Now data are readed by json.loads. Update test to pass builds
+
+Add support for different POST type (eg. FASTA)
+
+Deal with POST methods and parameters, like:
+curl 'http://rest.ensembl.org/sequence/region/human?expand_3prime=100' -H 'Content-type:application/json' -H 'Accept:application/json' -X POST -d '{ "regions" : ["X:1000000..1000100:1", "ABBA01004489.1:1..100"]}'
+[{"id":"chromosome:GRCh38:X:1000000:1000200:1","seq":"CTGTAGAAACATTAGCCTGGCTAACAAGGTGAAACCCCATCTCTACTAACAATACAAAATATTGGTTGGGCGTGGTGGCGGGTGCTTGTAATCCCAGCTACTCAGGAGGCTGAGGCAAGAGAATTGGTTGAACCCAGGAGGCGGAGGTTGCATTGAGCTGAGATCACGCCACTGCACTCCAGCCTGGGCAACAGAGCGAGA","molecule":"dna"},{"id":"clone::ABBA01004489.1:1:200:1","seq":"CTGTACTTTCCTTGGGATGGAGTAGTTTCGAAACACACTTTCTGTAGAATCTGCAAGTGGATATTTGGACCTGTCTGAGGAATTCGTTGGAAACGGGATAACTTCAGCTAACTAAACAGAAGCAATCTCCGAATCTTCTTTGTGATGTTTGCATTCGAATCCCAGAATTGAACCTTCCTTTGAAAGTTCAGGTTTGAAAC","molecule":"dna"}]
