+++ ensemblrest/ensembl_config.py
@@ -1,6 +1,7 @@
 """
 
     This file is part of pyEnsemblRest.
+    Copyright (C) 2013-2016, Steve Moss
 
     pyEnsemblRest is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
